--- src/main/java/woowacourse/shoppingcart/domain/Product.java
@@ -1,39 +1,0 @@
-package woowacourse.shoppingcart.domain;
-
-public class Product {
-
-    private Long id;
-    private String name;
-    private Integer price;
-    private String imageUrl;
-
-    public Product() {
-    }
-
-    public Product(final Long id, final String name, final int price, final String imageUrl) {
-        this.id = id;
-        this.name = name;
-        this.price = price;
-        this.imageUrl = imageUrl;
-    }
-
-    public Product(final String name, final int price, final String imageUrl) {
-        this(null, name, price, imageUrl);
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public int getPrice() {
-        return price;
-    }
-
-    public String getImageUrl() {
-        return imageUrl;
-    }
-
-    public Long getId() {
-        return id;
-    }
-}
