+++ src/test/java/woowacourse/shoppingcart/domain/customer/CustomerTest.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.domain;
+package woowacourse.shoppingcart.domain.customer;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
